+++ fda-model/compute/MeanRandomProvider.cs
@@ -23,7 +23,7 @@ namespace compute
         }
         
 
-        public double[] NextRandomSequence(int size)
+        public double[] NextRandomSequence(Int64 size)
         {
             double[] randomNumbers = new double[size];//needs to be initialized with a set of random nubmers between 0 and 1;
             for (int i = 0; i < size; i++)
